refactor(api): move /api/admin/users/{id}/add-pix-authentication-method route in IAM context
