--- api/lib/domain/usecases/add-pix-authentication-method-by-email.js
@@ -1,37 +1,0 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../src/identity-access-management/domain/constants/identity-providers.js';
-import { AuthenticationMethod } from '../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
-import { AuthenticationMethodAlreadyExistsError } from '../../../src/shared/domain/errors.js';
-
-const addPixAuthenticationMethodByEmail = async function ({
-  userId,
-  email,
-  passwordGenerator,
-  cryptoService,
-  userRepository,
-  authenticationMethodRepository,
-}) {
-  await userRepository.checkIfEmailIsAvailable(email);
-
-  const alreadyHasPixAuthenticationMethod = await authenticationMethodRepository.hasIdentityProviderPIX({ userId });
-
-  if (alreadyHasPixAuthenticationMethod) {
-    throw new AuthenticationMethodAlreadyExistsError();
-  } else {
-    const generatedPassword = passwordGenerator.generateComplexPassword();
-    const hashedPassword = await cryptoService.hashPassword(generatedPassword);
-
-    const authenticationMethodFromPix = new AuthenticationMethod({
-      userId,
-      identityProvider: NON_OIDC_IDENTITY_PROVIDERS.PIX.code,
-      authenticationComplement: new AuthenticationMethod.PixAuthenticationComplement({
-        password: hashedPassword,
-        shouldChangePassword: false,
-      }),
-    });
-    await authenticationMethodRepository.create({ authenticationMethod: authenticationMethodFromPix });
-    await userRepository.updateUserDetailsForAdministration({ id: userId, userAttributes: { email } });
-    return userRepository.getUserDetailsForAdmin(userId);
-  }
-};
-
-export { addPixAuthenticationMethodByEmail };
