+++ api/tests/identity-access-management/acceptance/application/user/user.admin.route.test.js
@@ -415,6 +415,38 @@ describe('Acceptance | Identity Access Management | Application | Route | Admin 
     });
   });
 
+  describe('POST /api/users/{id}/add-pix-authentication-method', function () {
+    it('returns 201 HTTP status code', async function () {
+      // given
+      const server = await createServer();
+      const superAdmin = await insertUserWithRoleSuperAdmin();
+      const user = databaseBuilder.factory.buildUser({ email: null });
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject({
+        method: 'POST',
+        url: `/api/admin/users/${user.id}/add-pix-authentication-method`,
+        headers: {
+          authorization: generateValidRequestAuthorizationHeader(superAdmin.id),
+        },
+        payload: {
+          data: {
+            id: user.id,
+            attributes: {
+              email: 'user@example.net',
+            },
+          },
+        },
+      });
+
+      // then
+      expect(response.statusCode).to.equal(201);
+      expect(response.result.data.attributes.email).to.equal('user@example.net');
+      expect(response.result.included[0].attributes['identity-provider']).to.equal('PIX');
+    });
+  });
+
   describe('POST /api/admin/users/{id}/remove-authentication', function () {
     let server;
     let user;
