+++ api/tests/unit/application/users/index_test.js
@@ -49,107 +49,6 @@ describe('Unit | Router | user-router', function () {
   });
 
   context('Routes /admin', function () {
-    describe('POST /api/admin/users/{id}/add-pix-authentication-method', function () {
-      it('returns 200 when user role is "SUPER_ADMIN"', async function () {
-        // given
-        sinon
-          .stub(userController, 'addPixAuthenticationMethodByEmail')
-          .callsFake((request, h) => h.response({}).code(201));
-        sinon
-          .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-          .callsFake((request, h) => h.response(true));
-        sinon
-          .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-          .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-        const payload = { data: { attributes: { email: 'user@rexample.net' } } };
-
-        // when
-        const { statusCode } = await httpTestServer.request(
-          'POST',
-          '/api/admin/users/1/add-pix-authentication-method',
-          payload,
-        );
-
-        // then
-        expect(statusCode).to.equal(201);
-        sinon.assert.calledOnce(securityPreHandlers.checkAdminMemberHasRoleSuperAdmin);
-        sinon.assert.calledOnce(securityPreHandlers.checkAdminMemberHasRoleSupport);
-        sinon.assert.calledOnce(userController.addPixAuthenticationMethodByEmail);
-      });
-
-      it('returns 200 when user role is "SUPPORT"', async function () {
-        // given
-        sinon
-          .stub(userController, 'addPixAuthenticationMethodByEmail')
-          .callsFake((request, h) => h.response({}).code(201));
-        sinon
-          .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-          .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-        sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport').callsFake((request, h) => h.response(true));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-        const payload = { data: { attributes: { email: 'user@rexample.net' } } };
-
-        // when
-        const { statusCode } = await httpTestServer.request(
-          'POST',
-          '/api/admin/users/1/add-pix-authentication-method',
-          payload,
-        );
-
-        // then
-        expect(statusCode).to.equal(201);
-        sinon.assert.calledOnce(securityPreHandlers.checkAdminMemberHasRoleSuperAdmin);
-        sinon.assert.calledOnce(securityPreHandlers.checkAdminMemberHasRoleSupport);
-        sinon.assert.calledOnce(userController.addPixAuthenticationMethodByEmail);
-      });
-
-      it('returns 400 when id is not a number', async function () {
-        // given
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const { statusCode, payload } = await httpTestServer.request(
-          'POST',
-          '/api/admin/users/invalid-id/add-pix-authentication-method',
-        );
-
-        // then
-        expect(statusCode).to.equal(400);
-        expect(JSON.parse(payload).errors[0].detail).to.equal('"id" must be a number');
-      });
-
-      it(`returns 403 when user don't have access (CERTIF | METIER)`, async function () {
-        // given
-        sinon.stub(userController, 'addPixAuthenticationMethodByEmail').returns('ok');
-        sinon
-          .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-          .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-        sinon
-          .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-          .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-        const payload = { data: { attributes: { email: 'user@rexample.net' } } };
-
-        // when
-        const result = await httpTestServer.request(
-          'POST',
-          '/api/admin/users/1/add-pix-authentication-method',
-          payload,
-        );
-
-        // then
-        expect(result.statusCode).to.equal(403);
-        sinon.assert.calledOnce(securityPreHandlers.checkAdminMemberHasRoleSuperAdmin);
-        sinon.assert.calledOnce(securityPreHandlers.checkAdminMemberHasRoleSupport);
-        sinon.assert.notCalled(userController.addPixAuthenticationMethodByEmail);
-      });
-    });
-
     describe('POST /api/admin/users/{userId}/authentication-methods/{authenticationMethodId}', function () {
       // eslint-disable-next-line mocha/no-setup-in-describe
       [CODE_IDENTITY_PROVIDER_GAR, CODE_IDENTITY_PROVIDER_POLE_EMPLOI, oidcProviderCode].forEach((identityProvider) => {
